requirements.txt
asgiref==3.7.2
Django==5.0.1
django-bootstrap-modal-forms==3.0.4
django-ckeditor==6.7.0
django-filebrowser==4.0.3
django-grappelli==3.0.8
django-js-asset==2.2.0
django-tinymce==3.6.1
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
